requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
propcache==0.3.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
twilio==9.4.6
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
gunicorn>=20.1.0
firebase-admin==6.2.0